+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Add JsonTags.manage to specify tags for a given JSON structure
+
 ## [3.0.16] - 2026-03-06
 
 ### Changed
@@ -115,6 +121,7 @@
 - Add `DeepCopy(json)`
 - Add deepCopy, deepEquals and isJson and json
 
+[Unreleased]: https://github.com/inlavigo/gg_json/compare/3.0.16...HEAD
 [3.0.16]: https://github.com/inlavigo/gg_json/compare/3.0.15...3.0.16
 [3.0.15]: https://github.com/inlavigo/gg_json/compare/3.0.14...3.0.15
 [3.0.14]: https://github.com/inlavigo/gg_json/compare/3.0.13...3.0.14